Introduce ActionPointsAdjustedRule. (#59)
* Introduce ActionPointsAdjustedRule.

This rule adjusts the starting action points for any specified PieceConfig.

Add documentation describing expected values of arguments to ActionPointsAdjustedRule.

* Switch to applying rule effects from OnActivate() to OnPostGameCreated().

* Update method name.

* Delete now-uneeded ballista-specific AP rule.
