+++ Rules/README.md
@@ -9,7 +9,7 @@ RulesAPI.
 
 - **SampleRule**: A [sample rule](Rule/SampleRule.cs) documenting the anatomy
   of a RulesAPI rule.
-- **BallistaActionPointsAdjustedRule**: Ballista action points are adjusted
+- **ActionPointsAdjustedRule**: Action points are adjusted
 - **BallistaAttackDamageAdjustedRule**: Ballista attack damage is adjusted
 - **RatNestsSpawnGoldRule**: Rat nests spawn gold
 - **ZapDamageAdjustedRule**: Zap damage is adjusted
