+++ Rules/Rules.csproj
@@ -34,7 +34,7 @@
     <ItemGroup>
         <Compile Include="Properties\AssemblyInfo.cs" />
         <Compile Include="RulesMod.cs" />
-        <Compile Include="Rule\BallistaActionPointsAdjustedRule.cs" />
+        <Compile Include="Rule\ActionPointsAdjustedRule.cs" />
         <Compile Include="Rule\BallistaAttackDamageAdjustedRule.cs" />
         <Compile Include="Rule\RatNestsSpawnGoldRule.cs" />
         <Compile Include="Rule\SampleRule.cs" />
