+++ Rules/RulesMod.cs
@@ -5,6 +5,8 @@
 
     internal class RulesMod : MelonMod
     {
+        internal static readonly MelonLogger.Instance Logger = new MelonLogger.Instance("Rules");
+
         public override void OnApplicationStart()
         {
             RegisterNewRuleTypes();
@@ -15,7 +17,7 @@
         {
             var registrar = RulesAPI.Registrar.Instance();
             registrar.Register(typeof(Rule.SampleRule));
-            registrar.Register(typeof(Rule.BallistaActionPointsAdjustedRule));
+            registrar.Register(typeof(Rule.ActionPointsAdjustedRule));
             registrar.Register(typeof(Rule.BallistaAttackDamageAdjustedRule));
             registrar.Register(typeof(Rule.RatNestsSpawnGoldRule));
             registrar.Register(typeof(Rule.ZapDamageAdjustedRule));
